整理加速结构

+++ JustLab/Accel/AccelNode.h
@@ -6,15 +6,15 @@
 
 namespace Just {
     struct AccelNode {
-        int child;
+        size_t child;
         Bounds3f bbox;
-        std::vector<std::pair<int, int>> indexes;
+        std::vector<std::pair<size_t, size_t>> indexes;
 
         AccelNode() : bbox(), child(0) {}
 
         explicit AccelNode(const Bounds3f& bbox) : bbox(bbox), child(0) {}
 
-        AccelNode(const Bounds3f& bbox, int size)
+        AccelNode(const Bounds3f& bbox, size_t size)
                 : bbox(bbox), indexes(size), child(0) {}
     };
 }
